Add an efficient enumerator impl to EntityQuery.
