+++ Robust.Shared/GameObjects/EntityManager.Components.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Frozen;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
@@ -1188,16 +1187,21 @@ namespace Robust.Shared.GameObjects
 
         public EntityQuery<TComp1> GetEntityQuery<TComp1>() where TComp1 : IComponent
         {
+            DebugTools.Assert(_entTraitArray.Length > CompIdx.ArrayIndex<TComp1>(),
+                $"Unknown component: {typeof(TComp1).Name}");
             var comps = _entTraitArray[CompIdx.ArrayIndex<TComp1>()];
-            DebugTools.Assert(comps != null, $"Unknown component: {typeof(TComp1).Name}");
-            return new EntityQuery<TComp1>(this, comps);
+            var meta = _entTraitArray[CompIdx.ArrayIndex<MetaDataComponent>()];
+
+            return new EntityQuery<TComp1>(this, comps, meta);
         }
 
         public EntityQuery<IComponent> GetEntityQuery(Type type)
         {
+            DebugTools.Assert(_entTraitDict.ContainsKey(type), $"Unknown component: {type.Name}");
             var comps = _entTraitDict[type];
-            DebugTools.Assert(comps != null, $"Unknown component: {type.Name}");
-            return new EntityQuery<IComponent>(this, comps);
+            var meta = _entTraitArray[CompIdx.ArrayIndex<MetaDataComponent>()];
+
+            return new EntityQuery<IComponent>(this, comps, meta);
         }
 
         /// <inheritdoc />
@@ -1763,337 +1767,6 @@ namespace Robust.Shared.GameObjects
                 return (val.Key, val.Value);
             }
         }
-    }
-
-    /// <summary>
-    ///     An index of all entities with a given component, avoiding looking up the component's storage every time.
-    ///     Using these saves on dictionary lookups, making your code slightly more efficient, and ties in nicely with
-    ///     <see cref="Entity{T}"/>.
-    /// </summary>
-    /// <typeparam name="TComp1">Any component type.</typeparam>
-    /// <example>
-    ///     <code>
-    ///         public sealed class MySystem : EntitySystem
-    ///         {
-    ///             private EntityQuery&lt;TransformComponent&gt; _transforms = default!;
-    ///             <br/>
-    ///             public override void Initialize()
-    ///             {
-    ///                 _transforms = GetEntityQuery&lt;TransformComponent&gt;();
-    ///             }
-    ///             <br/>
-    ///             public void DoThings(EntityUid myEnt)
-    ///             {
-    ///                 var ent = _transforms.Get(myEnt);
-    ///                 // ...
-    ///             }
-    ///         }
-    ///     </code>
-    /// </example>
-    /// <remarks>
-    ///     Queries hold references to <see cref="IEntityManager"/> internals, and are always up to date with the world.
-    ///     They can not however perform mutation, if you need to add or remove components you must use
-    ///     <see cref="EntitySystem"/> or <see cref="IEntityManager"/> methods.
-    /// </remarks>
-    /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.GetEntityQuery``1">EntitySystem.GetEntityQuery()</seealso>
-    /// <seealso cref="M:Robust.Shared.GameObjects.EntityManager.GetEntityQuery``1">EntityManager.GetEntityQuery()</seealso>
-    public readonly struct EntityQuery<TComp1> where TComp1 : IComponent
-    {
-        private readonly EntityManager _entMan;
-        private readonly Dictionary<EntityUid, IComponent> _traitDict;
-
-        internal EntityQuery(EntityManager entMan, Dictionary<EntityUid, IComponent> traitDict)
-        {
-            _entMan = entMan;
-            _traitDict = traitDict;
-        }
-
-        /// <summary>
-        ///     Gets <typeparamref name="TComp1"/> for an entity, throwing if it can't find it.
-        /// </summary>
-        /// <param name="uid">The entity to do a lookup for.</param>
-        /// <returns>The located component.</returns>
-        /// <exception cref="KeyNotFoundException">Thrown if the entity does not have a component of type <typeparamref name="TComp1"/>.</exception>
-        /// <seealso cref="M:Robust.Shared.GameObjects.IEntityManager.GetComponent``1(Robust.Shared.GameObjects.EntityUid)">
-        ///     IEntityManager.GetComponent&lt;T&gt;(EntityUid)
-        /// </seealso>
-        /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.Comp``1(Robust.Shared.GameObjects.EntityUid)">
-        ///     EntitySystem.Comp&lt;T&gt;(EntityUid)
-        /// </seealso>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public TComp1 GetComponent(EntityUid uid)
-        {
-            if (_traitDict.TryGetValue(uid, out var comp) && !comp.Deleted)
-                return (TComp1) comp;
-
-            throw new KeyNotFoundException($"Entity {uid} does not have a component of type {typeof(TComp1)}");
-        }
-
-        /// <inheritdoc cref="GetComponent"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
-        public Entity<TComp1> Get(EntityUid uid)
-        {
-            if (_traitDict.TryGetValue(uid, out var comp) && !comp.Deleted)
-                return new Entity<TComp1>(uid, (TComp1) comp);
-
-            throw new KeyNotFoundException($"Entity {uid} does not have a component of type {typeof(TComp1)}");
-        }
-
-        /// <summary>
-        ///     Gets <typeparamref name="TComp1"/> for an entity, if it's present.
-        /// </summary>
-        /// <remarks>
-        ///     If it is strictly errorenous for a component to not be present, you may want to use
-        ///     <see cref="Resolve(Robust.Shared.GameObjects.EntityUid,ref TComp1?,bool)"/> instead.
-        /// </remarks>
-        /// <param name="uid">The entity to do a lookup for.</param>
-        /// <param name="component">The located component, if any.</param>
-        /// <returns>Whether the component was found.</returns>
-        /// <seealso cref="M:Robust.Shared.GameObjects.IEntityManager.TryGetComponent``1(Robust.Shared.GameObjects.EntityUid,``0@)">
-        ///     IEntityManager.TryGetComponent&lt;T&gt;(EntityUid, out T?)
-        /// </seealso>
-        /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.TryComp``1(Robust.Shared.GameObjects.EntityUid,``0@)">
-        ///     EntitySystem.TryComp&lt;T&gt;(EntityUid, out T?)
-        /// </seealso>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool TryGetComponent([NotNullWhen(true)] EntityUid? uid, [NotNullWhen(true)] out TComp1? component)
-        {
-            if (uid == null)
-            {
-                component = default;
-                return false;
-            }
-
-            return TryGetComponent(uid.Value, out component);
-        }
-
-        /// <inheritdoc cref="TryGetComponent(Robust.Shared.GameObjects.EntityUid?,out TComp1?)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool TryGetComponent(EntityUid uid, [NotNullWhen(true)] out TComp1? component)
-        {
-            if (_traitDict.TryGetValue(uid, out var comp) && !comp.Deleted)
-            {
-                component = (TComp1) comp;
-                return true;
-            }
-
-            component = default;
-            return false;
-        }
-
-        /// <inheritdoc cref="TryGetComponent(Robust.Shared.GameObjects.EntityUid?,out TComp1?)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool TryComp(EntityUid uid, [NotNullWhen(true)] out TComp1? component)
-            => TryGetComponent(uid, out component);
-
-        /// <inheritdoc cref="TryGetComponent(Robust.Shared.GameObjects.EntityUid?,out TComp1?)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool TryComp([NotNullWhen(true)] EntityUid? uid, [NotNullWhen(true)] out TComp1? component)
-            => TryGetComponent(uid, out component);
-
-        /// <summary>
-        ///     Tests if the given entity has <typeparamref name="TComp1"/>.
-        /// </summary>
-        /// <param name="uid">The entity to do a lookup for.</param>
-        /// <returns>Whether the component exists for that entity.</returns>
-        /// <remarks>If you immediately need to then look up that component, it's more efficient to use <see cref="TryComp(Robust.Shared.GameObjects.EntityUid,out TComp1?)"/>.</remarks>
-        /// <seealso cref="M:Robust.Shared.GameObjects.IEntityManager.HasComponent``1(Robust.Shared.GameObjects.EntityUid)">
-        ///     IEntityManager.HasComponent&lt;T&gt;(EntityUid)
-        /// </seealso>
-        /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.HasComp``1(Robust.Shared.GameObjects.EntityUid)">
-        ///     EntitySystem.HasComp&lt;T&gt;(EntityUid)
-        /// </seealso>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool HasComp(EntityUid uid) => HasComponent(uid);
-
-        /// <inheritdoc cref="HasComp(Robust.Shared.GameObjects.EntityUid)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool HasComp([NotNullWhen(true)] EntityUid? uid) => HasComponent(uid);
-
-        /// <inheritdoc cref="HasComp(Robust.Shared.GameObjects.EntityUid)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool HasComponent(EntityUid uid)
-        {
-            return _traitDict.TryGetValue(uid, out var comp) && !comp.Deleted;
-        }
-
-        /// <inheritdoc cref="HasComp(Robust.Shared.GameObjects.EntityUid)"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public bool HasComponent([NotNullWhen(true)] EntityUid? uid)
-        {
-            return uid != null && HasComponent(uid.Value);
-        }
-
-        /// <include file='Docs.xml' path='entries/entry[@name="EntityQueryResolve"]/*'/>
-        /// <param name="uid">The entity to do a lookup for.</param>
-        /// <param name="component">The space to write the component into if found.</param>
-        /// <param name="logMissing">Whether to log if the component is missing, for diagnostics.</param>
-        /// <returns>Whether the component was found.</returns>
-        /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.Resolve``1(Robust.Shared.GameObjects.EntityUid,``0@,System.Boolean)">
-        ///     EntitySystem.Resolve&lt;T&gt;(EntityUid, out T?)
-        /// </seealso>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public bool Resolve(EntityUid uid, [NotNullWhen(true)] ref TComp1? component, bool logMissing = true)
-        {
-            if (component != null)
-            {
-                DebugTools.AssertOwner(uid, component);
-                return true;
-            }
-
-            if (_traitDict.TryGetValue(uid, out var comp) && !comp.Deleted)
-            {
-                component = (TComp1)comp;
-                return true;
-            }
-
-            if (logMissing)
-                _entMan.ResolveSawmill.Error($"Can't resolve \"{typeof(TComp1)}\" on entity {_entMan.ToPrettyString(uid)}!\n{Environment.StackTrace}");
-
-            return false;
-        }
-
-        /// <include file='Docs.xml' path='entries/entry[@name="EntityQueryResolve"]/*'/>
-        /// <param name="entity">The space to write the component into if found.</param>
-        /// <param name="logMissing">Whether to log if the component is missing, for diagnostics.</param>
-        /// <returns>Whether the component was found.</returns>
-        /// <seealso cref="M:Robust.Shared.GameObjects.EntitySystem.Resolve``1(Robust.Shared.GameObjects.EntityUid,``0@,System.Boolean)">
-        ///     EntitySystem.Resolve&lt;T&gt;(EntityUid, out T?)
-        /// </seealso>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public bool Resolve(ref Entity<TComp1?> entity, bool logMissing = true)
-        {
-            return Resolve(entity.Owner, ref entity.Comp, logMissing);
-        }
-
-        /// <summary>
-        ///     Gets <typeparamref name="TComp1"/> for an entity if it's present, or null if it's not.
-        /// </summary>
-        /// <param name="uid">The entity to do the lookup on.</param>
-        /// <returns>The component, if it exists.</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public TComp1? CompOrNull(EntityUid uid)
-        {
-            if (TryGetComponent(uid, out var comp))
-                return comp;
-
-            return default;
-        }
-
-        /// <inheritdoc cref="GetComponent"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        public TComp1 Comp(EntityUid uid)
-        {
-            return GetComponent(uid);
-        }
-
-        #region Internal
-
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="GetComponent"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal TComp1 GetComponentInternal(EntityUid uid)
-        {
-            if (_traitDict.TryGetValue(uid, out var comp))
-                return (TComp1) comp;
-
-            throw new KeyNotFoundException($"Entity {uid} does not have a component of type {typeof(TComp1)}");
-        }
-
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="TryGetComponent(System.Nullable{Robust.Shared.GameObjects.EntityUid},out TComp1?)"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal bool TryGetComponentInternal([NotNullWhen(true)] EntityUid? uid, [NotNullWhen(true)] out TComp1? component)
-        {
-            if (uid == null)
-            {
-                component = default;
-                return false;
-            }
-
-            return TryGetComponentInternal(uid.Value, out component);
-        }
-
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="TryGetComponent(System.Nullable{Robust.Shared.GameObjects.EntityUid},out TComp1?)"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal bool TryGetComponentInternal(EntityUid uid, [NotNullWhen(true)] out TComp1? component)
-        {
-            if (_traitDict.TryGetValue(uid, out var comp))
-            {
-                component = (TComp1) comp;
-                return true;
-            }
-
-            component = default;
-            return false;
-        }
-
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="HasComponent(EntityUid)"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal bool HasComponentInternal(EntityUid uid)
-        {
-            return _traitDict.TryGetValue(uid, out var comp) && !comp.Deleted;
-        }
-
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="Resolve"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal bool ResolveInternal(EntityUid uid, [NotNullWhen(true)] ref TComp1? component, bool logMissing = true)
-        {
-            if (component != null)
-            {
-                DebugTools.AssertOwner(uid, component);
-                return true;
-            }
-
-            if (_traitDict.TryGetValue(uid, out var comp))
-            {
-                component = (TComp1)comp;
-                return true;
-            }
-
-            if (logMissing)
-                _entMan.ResolveSawmill.Error($"Can't resolve \"{typeof(TComp1)}\" on entity {_entMan.ToPrettyString(uid)}!\n{new StackTrace(1, true)}");
-
-            return false;
-        }
-        /// <summary>
-        /// Elides the component.Deleted check of <see cref="CompOrNull"/>
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [Pure]
-        internal TComp1? CompOrNullInternal(EntityUid uid)
-        {
-            if (TryGetComponent(uid, out var comp))
-                return comp;
-
-            return default;
-        }
-
-        #endregion
     }
 
     #region ComponentRegistry Query
